- Compile in XCode 9.1 - Deployment target increased up to iOS 8.0 - Get rid of most compilation warnings - Facebook SDK updated

+++ Classes/FriendRequestsProfilesController.m
@@ -76,7 +76,7 @@
 	
 	[indexPathForRemoving release];
 	indexPathForRemoving = [[table indexPathForCell:cell] retain];
-	NSLog(@"%@ at row: %d", aMethod, indexPathForRemoving.row);	
+	NSLog(@"%@ at row: %d", aMethod, (int)indexPathForRemoving.row);	
 }
 
 /**********************************************************************************************************************
